Create ElectLeader
Details the most abstracted and thus most efficient version of the Elect Leader Protocol. Written using Alloy 6. 

Creates a closed ring of Node objects and allows them to send messages to each other in order to determine the leader via an inherent comparable property. When the leader is elected, all Nodes in the ring should acknowledge this leader. 

Adapted from the Practical Alloy textbook.