Create HardwareScale
This Alloy model attempts to highlight a potential inconsistency in the described functional requirements specification for the O-RAN O2 interface. That is, we assume that no two clouds can own the same node, however, the sequence description of adding a new node to a cloud potentially allows such a state.