Update NF_lifecycle_infected_hardware
Update to description. 

The O-RAN Spec does describe functionality to heal faulty nodes, however, deployment flow does not describe what happens if a request is sent for an NF to be deployed upon a node that was initially functional, but has since become inoperable. 

It is unclear whether the request should wait for healing to occur or whether the deployment should be completed anyway. No exception implies that the request is not cancelled. 

This model shows the possibility for a NF to be deployed upon a currently faulty node.