Create HardwareScale_Semaphore
Here we implement a semaphore flag to solve race conditions. However, as observed in the checking of the given assertion, it is still possible for a single node to be deployed into two clouds. 

It is now clear that if this state is unwanted, a precondition must be added to ensure that a node is `uninitialized' before hardware scaling can occur.