Create HardwareScale_infected_cloud
This Alloy model shows the possibility that a process of adding a new node can be initiated but the cloud since becomes inoperable. No exceptions in the spec implies that the scaling either completes (violating preconditions), or the node waits until the cloud is healed. What if such healing is unsuccessful, does this merit an exception of some description?