Create VLAN_ID_allocation_duplicates
This model describes the sequence of requesting a VLAN ID from an O-Cloud instance (3.9.1). 
A VLAN ID is requested as part of a network slice initialization. This sequence is repeated for each cloud part of the slice.
It states that each VLAN ID should only be distributed once per cloud, as they hold local significance.
This Alloy model highlights an instance where two clouds, which both hold the same ID, receive a slice request. 
Here, both clouds respond with the same ID. 
It is not strictly forbidden in the spec for such a scenario to occur. 
From research, this scenario seems to be acceptable, but can cause trouble if an implementation does not account for it.