Create VLAN_lifecycle
This model describes the life cycles of VLAN IDs. 

When slices are made, clouds involved send away an ID that is not currently used. 

When the slice is destroyed the ID should be returned to the Cloud it came from. 

This model verifies that once a VLAN ID has been returned, that it able to be redeployed for a new request.