Create Create_Network_Slice_No_Deployed
In the sequence description of creating a new network slice (clause 3.10.1), it says that the a VLAN ID allocation request is sent to each O-Cloud involved in the slice. 
It later gives the option for additional NFs to be deployed and given the selected ID. 

What happens in the case where a slice request which contains no deployed NFs is initiated? How does the system know what O-Cloud to target for VLAN allocation? 

In the case described in this model, if a request is generated and only contains un-deployed NFs, then the system reaches a deadlock where VLAN ID allocation is never attempted. 

To amend this we either require that at least one requested NF is already deployed, or functionality for homing an appropriate O-Cloud for VLAN Allocation is implemented.