Create Network_Slice_Creation
//WORK IN PROGRESS//

Adapted from clause 3.10.1 of the O-RAN O2 Use case Spec. 

This details the creation of a new network slice. This is done by requesting a VLAN ID from each involved cloud that will be attached to each NF in that clouds portion of the slice. There is also functionallity to deploy additional NFs. If the slice does span multiple clouds, then connectivity testing must be deployed#.