Create Node_Cluster_Creation
This details the creation of a node cluster which provisions the nodes inside with functionality to deploy NF. Some basic assertions are verified such as;

- Clusters are non-empty
- Clusters are disjoint
- ID associated w/ provisioned resource matches that of the cluster
- Two identical requests sent at the same time will cause one to fail and one to succeed
- IDs are locally unique
- Booked resources are put into the correct cluster
- Resources in different clusters, but the same cloud, have different cluster IDs

Building upon the possible ambiguity in clause 3.1.5 - Hardware Scaling, there may be an unwanted state possible. That is, of one node can be shared among multiple clouds, then who is handled when cluster creation requests are made? This model says that this leads to cases where an O-Cloud has a node in its inventory that is provisioned, but not by a cluster directly within that O-Cloud. 

This may lead to further situations where NFs are deployed upon a un-provisioned node from the perspective of an O-Cloud.